commit not to push

+++ src/pages/Gameroom.jsx
@@ -5,8 +5,12 @@ import { boardsService } from "../services/boards-service"
 import { boardService } from "../services/board-service"
 import { socketService } from "../services/socketService";
 
-import { GameData } from "../cmps/GameData.jsx";
+// import { GameData } from "../cmps/GameData.jsx";
 import { useSelector } from "react-redux"
+
+import {
+  refresh,
+} from '../store/actions/userActions'
 
 export const Gameroom = () => {
   const params = useParams();
@@ -27,8 +31,10 @@ export const Gameroom = () => {
   }
 
   const onLeavingGame = async () => {
-    if (user.loggedInUser && (user.loggedInUser._id === game.whitePlayer.user._id || user.loggedInUser._id === game.blackPlayer.user._id)) {
+    if (user.loggedInUser && (user.loggedInUser._id === game.whitePlayer.user._id || (game.blackPlayer.user && user.loggedInUser._id === game.blackPlayer.user._id))) {
       await boardsService.deleteBoard(id)
+      socketService.emit('game-ended', true)
+      refresh(user.loggedInUser._id)
       console.log('game deleted');
     }
     history.goBack();
@@ -36,15 +42,24 @@ export const Gameroom = () => {
   useEffect(() => {
     async function socketSetup() {
       await socketService.setup();
-      socketService.emit("room-id", id);
-      socketService.on("update-board", updateSocketBoard);
+      socketService.emit('room-id', id);
+      socketService.on('update-board', updateSocketBoard);
+      socketService.on('game-ended', onEndGame)
+      socketService.emit('update-board', game);
     }
     socketSetup()
+  // eslint-disable-next-line
   }, [id])
 
   const updateSocketBoard = (game) => {
     setGame(game);
   };
+
+  const onEndGame = (bool) => {
+    console.log('end game')
+    refresh(user.loggedInUser._id)
+    bool && history.push('/gamerooms')
+  }
 
   useEffect(() => {
     if (!game) return
@@ -73,25 +88,30 @@ export const Gameroom = () => {
       setBoard(_board);
     }
     updateGameBoard();
-  }, [game,user.loggedInUser]);
+  }, [game, user.loggedInUser]);
 
-  //TODO: create cell component
-  // redux hold game, and user on the reducer make a function hold and remove from cell Selected..
   if (!game) {
     reset()
     return <div>loading...</div>
   }
-
   return (
     <div>
-      <h1>GameRoom {id}</h1>
-      <section className="flex wrap">
-        <div className={`board${(user.loggedInUser && game.blackPlayer.user && user.loggedInUser._id === game.blackPlayer.user._id) ? ' black-player' : ''}`}>
-          {board}
-        </div>
+      <h1>GameRoom {id} <button className="leave-game-btn" onClick={() => onLeavingGame()}>Leave Game</button></h1>
 
-        {game.blackPlayer.user ? <GameData game={game} /> : 'waiting for an opponent'}
-        <button onClick={() => onLeavingGame()}>Leave Game</button>
+      <section className="flex wrap justify-center">
+        <div className="main-board flex column">
+          <div className="opponent w-100">
+            {(user.loggedInUser && game.blackPlayer.user) ?
+              ((user.loggedInUser._id === game.blackPlayer.user._id) ? game.whitePlayer.user.username : game.blackPlayer.user.username) :
+              'Pending for another player...'}
+          </div>
+          <div className={`board${(user.loggedInUser && game.blackPlayer.user && user.loggedInUser._id === game.blackPlayer.user._id) ? ' black-player' : ''}`}>
+            {board}
+          </div>
+          <div className="opponent w-100">{(user.loggedInUser && game.blackPlayer.user && (user.loggedInUser._id === game.blackPlayer.user._id)) ? game.blackPlayer.user.username : game.whitePlayer.user.username}</div>
+        </div>
+        {/* Debbuging */}
+        {/* {game.blackPlayer.user ? <GameData game={game} /> : 'waiting for an opponent'} */}
       </section>
     </div>
   )
